chore(lights-controller) Split into modules.
